# Do not take snapshots when numTestsKeptInMemory is 0