--- a/packages/driver/src/cypress/log.coffee
+++ b/packages/driver/src/cypress/log.coffee
@@ -263,7 +263,7 @@ Log = (state, config, obj) ->
       ## bail early and dont snapshot
       ## if we're in headless mode
       ## TODO: fix this
-      if not config("isInteractive")
+      if not config("isInteractive") || _.toString(config("numTestsKeptInMemory")) is '0'
         return @
 
       _.defaults options,