closes #4104

Updated `log.coffee` to avoid snapshots when `numTestsKeptInMemory` is `0` in the same way that is done when `config("isInteractive")` is `true`. This resulted in a sample test with 50 assertions against an HTML document with a 10,000-rule CSS file, to run 40% faster. This also allows users running with `numTestsKeptInMemory` set to 0 to avoid the memory issues being addressed in #4068 

| | v3.20 | this branch |
| ---- | ---- | --- |
| 1 | 10.16s | 6.06s |
| 2 | 10.16s | 6.05s |
| 3 | 10.15s | 6.14s |
| 4 | 9.99s | 5.83s |
| 5 | 9.96s | 6.38s |
| avg | 10.08s | 6.09s |

In `packages/driver/test/support/server.coffee` I've also added a route that will generate a dynamically sized CSS file (per @brian-mann suggestion in #4068)

```coffeescript
  ## <link href="/dynamically-sized-css-file/:ruleCount.css" /> will include a stylesheet with 1000 rules
  app.get "/dynamically-sized-css-file/:ruleCount.css", (req, res) ->
```

----

## Screenshots

<img width="100%" style="margin: 0 auto;" alt="Screenshot of test run with numTestsKeptInMemory 0" src="https://user-images.githubusercontent.com/1461792/57168822-92ccd280-6dd1-11e9-90be-366ea00dbb8d.png">

In these 2 screenshots you can see that the most time spent is within different set of methods before and after these changes.
### Before
<img width="625" alt="Screen Shot 2019-05-03 at 6 48 38 PM" src="https://user-images.githubusercontent.com/1461792/57169397-66668580-6dd4-11e9-9602-255532740944.png">

### After
<img width="625" alt="Screen Shot 2019-05-03 at 6 50 19 PM" src="https://user-images.githubusercontent.com/1461792/57169399-66668580-6dd4-11e9-9251-dbd392c6e4dd.png">