@CoryDanielson 2019-05-04
Within my project for work, I saw a 58% decrease in run time for some login tests. We have ~20-35 requests per page load, and >10k CSS rules from external stylesheets on each page. The run time decreased from ~45s to ~19s, which was faster than the run time for `cypress run` with those same tests (24s).