@brian-mann 2019-05-05
This is a win. There are some failing tests that need to be fixed - because the tests are testing that snapshots happen. They need to be updated and likely need to introduce a "force snapshot anyway" so that in run mode we can test the behavior of snapshots (even if they would otherwise not happen).